# ModeLink/backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ModeLink API",
    "version": "0.1.0",
    "description": "Backend API для проекта ModeLink — цифровая экосистема управления моделями.\n\nАвторизация: Bearer JWT (header Authorization: Bearer <token>)."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {}
        },
        "required": [
          "error",
          "message"
        ]
      },
      "UserPublic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "model",
              "booker",
              "manager",
              "admin"
            ]
          },
          "isVerified": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "isVerified"
        ]
      },
      "AuthTokenResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserPublic"
          }
        },
        "required": [
          "token",
          "user"
        ]
      },
      "OtpRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "register"
            ],
            "default": "register"
          }
        },
        "required": [
          "email"
        ]
      },
      "OtpRequestResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "ok",
          "expiresAt"
        ]
      },
      "OtpVerify": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "register"
            ],
            "default": "register"
          },
          "code": {
            "type": "string",
            "pattern": "^\\\\d{6}$"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "role": {
            "type": "string",
            "enum": [
              "model",
              "booker",
              "manager",
              "admin"
            ]
          },
          "adminRegistrationSecret": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "code",
          "password",
          "role"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ModelProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "height": {
            "type": "integer",
            "nullable": true
          },
          "weight": {
            "type": "integer",
            "nullable": true
          },
          "bust": {
            "type": "integer",
            "nullable": true
          },
          "waist": {
            "type": "integer",
            "nullable": true
          },
          "hips": {
            "type": "integer",
            "nullable": true
          },
          "shoeSize": {
            "type": "number",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "fullName",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "AdminUserModelProfileSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "fullName"
        ]
      },
      "AdminUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "model",
              "booker",
              "manager",
              "admin"
            ]
          },
          "isVerified": {
            "type": "boolean"
          },
          "isBlocked": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modelProfile": {
            "$ref": "#/components/schemas/AdminUserModelProfileSummary"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "isVerified",
          "isBlocked",
          "createdAt"
        ]
      },
      "PagedAdminUsers": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminUser"
            }
          },
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 20
          },
          "total": {
            "type": "integer",
            "example": 100
          },
          "totalPages": {
            "type": "integer",
            "example": 5
          }
        },
        "required": [
          "items",
          "page",
          "limit",
          "total",
          "totalPages"
        ]
      },
      "AdminUserPatchRequest": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "model",
              "booker",
              "manager",
              "admin"
            ]
          },
          "isBlocked": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    }
  },
  "paths": {
...
    "/auth/otp/request": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Запросить OTP на email (регистрация)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP создан и отправлен (или залогирован в dev)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpRequestResponse"
                }
              }
            }
          },
          "409": {
            "description": "Пользователь уже существует",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Слишком часто запрашивают OTP (cooldown)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/otp/verify": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Подтвердить OTP + создать пользователя + задать пароль + выбрать роль",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerify"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно: возвращает JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Неверный/истёкший OTP или ошибка валидации",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Запрещена саморегистрация admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Логин по email + пароль",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно: возвращает JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Неверные креды",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Текущий пользователь",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "model",
                        "booker",
                        "manager",
                        "admin"
                      ]
                    },
                    "isVerified": {
                      "type": "boolean"
                    },
                    "modelProfile": {
                      "$ref": "#/components/schemas/ModelProfile"
                    }
                  },
                  "required": [
                    "id",
                    "email",
                    "role",
                    "isVerified"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Нет/неверный JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List users (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "required": false
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            },
            "required": false
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "email",
                "role"
              ]
            },
            "required": false
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "required": false
          },
          {
            "name": "role",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "model",
                "booker",
                "manager",
                "admin"
              ]
            },
            "required": false
          },
          {
            "name": "blocked",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Paged users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedAdminUsers"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/admin/users/{id}": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "Update user (role/block) (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserPatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUser"
                }
              }
            }
          },
          "400": {
            "description": "BadRequest"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "NotFound"
          }
        }
      }
    }
  }
}